+++ avm_abi/src/abi_token.rs
@@ -1,4 +1,34 @@
 
+#![allow(unused)]
+
+use std::mem;
+
+pub trait ToBytes {
+    fn to_bytes(&self) -> Vec<u8>;
+}
+
+macro_rules! format_as_bytes {
+    ($type_name: ident, $endian: expr, $len: expr) => {
+        impl ToBytes for $type_name {
+            fn to_bytes(&self) -> Vec<u8> {
+                let bytes: [u8; $len] = match $endian {
+                    0 => unsafe {mem::transmute(self.to_be())},
+                    1 => unsafe {mem::transmute(self.to_le())},
+                    _ => [0; $len]
+                };
+                bytes.to_vec()
+            }
+        }
+    };
+}
+
+format_as_bytes!(u16, 0, 2);
+format_as_bytes!(i16, 0, 2);
+format_as_bytes!(u32, 0, 4);
+format_as_bytes!(i32, 0, 4);
+format_as_bytes!(u64, 0, 8);
+format_as_bytes!(i64, 0, 8);
+
 enum AbiToken<'a> {
     UCHAR(u8),
     BOOL(bool),
@@ -45,46 +75,79 @@ impl<'a> AVMEncoder for AbiToken<'a> {
             },
             AbiToken::INT16(v) => {
                 res.push(0x04);
-                res.append(v.to_bytes())
+                res.append(&mut v.to_bytes())
             },
             AbiToken::INT32(v) => {
                 res.push(0x05);
+                res.append(&mut v.to_bytes())
             },
             AbiToken::INT64(v) => {
                 res.push(0x06);
+                res.append(&mut v.to_bytes())
             },
             AbiToken::FLOAT(v) => {
                 res.push(0x07);
+                //TODO: float variable
+                panic!("unsupported for float");
             },
             AbiToken::DOUBLE(v) => {
                 res.push(0x08);
+                //TODO: double variable
+                panic!("unsupported for double");
             },
             AbiToken::AUCHAR(v) => {
                 res.push(0x11);
+                v.iter().map(|v| {
+                    res.push(*v)
+                });
             },
             AbiToken::ABOOL(v) => {
                 res.push(0x12);
+                v.iter().map(|v| {
+                    if *v {
+                        res.push(0x01)
+                    } else {
+                        res.push(0x02)
+                    }
+                });
             },
             AbiToken::AINT8(v) => {
                 res.push(0x13);
+                v.iter().map(|v| {
+                    res.push(*v as u8)
+                });
             },
             AbiToken::AINT16(v) => {
                 res.push(0x14);
+                v.iter().map(|v| {
+                    res.append(&mut v.to_bytes())
+                });
             },
             AbiToken::AINT32(v) => {
                 res.push(0x15);
+                v.iter().map(|v| {
+                    res.append(&mut v.to_bytes())
+                });
             },
             AbiToken::AINT64(v) => {
                 res.push(0x16);
+                v.iter().map(|v| {
+                    res.append(&mut v.to_bytes())
+                });
             },
             AbiToken::AFLOAT(v) => {
                 res.push(0x17);
+                //TODO: float array
+                panic!("unsupported for float array");
             },
             AbiToken::ADOUBLE(v) => {
-                res.push(0x18)
+                res.push(0x18);
+                panic!("unsupported for double array");
             },
-            AbiToken::STRING(v) => {
+            AbiToken::STRING(ref v) => {
                 res.push(0x21);
+                res.append(&mut (v.len() as i32).to_bytes());
+                res.append(&mut v.clone().into_bytes());
             },
         }
 
@@ -92,27 +155,14 @@ impl<'a> AVMEncoder for AbiToken<'a> {
     }
 }
 
+#[cfg(test)]
+mod tests {
+    use super::*;
 
-// impl TokenType {
-//     fn encode(&self) -> Option<u8> {
-//         match *self {
-//             TokenType::UCHAR => Some(0x01),
-//             TokenType::BOOL => Some(0x02),
-//             TokenType::INT8 => Some(0x03),
-//             TokenType::INT16 => Some(0x04),
-//             TokenType::INT32 => Some(0x05),
-//             TokenType::INT64 => Some(0x06),
-//             TokenType::FLOAT => Some(0x07),
-//             TokenType::DOUBLE => Some(0x08),
-//             TokenType::AUCHAR => Some(0x11),
-//             TokenType::ABOOL => Some(0x12),
-//             TokenType::AINT8 => Some(0x13),
-//             TokenType::AINT16 => Some(0x14),
-//             TokenType::AINT32 => Some(0x15),
-//             TokenType::AINT64 => Some(0x16),
-//             TokenType::AFLOAT => Some(0x17),
-//             TokenType::ADOUBLE => Some(0x18),
-//             TokenType::STRING => Some(0x21),
-//         }
-//     }
-// }
+    #[test]
+    fn encode() {
+        let token = AbiToken::STRING("hello".to_string());
+
+        assert_eq!(token.encode(), vec![33,0,0,0,5,104,101,108,108,111]);
+    }
+}
