+++ avm_abi/src/lib.rs
@@ -1,3 +1,3 @@
 
-mod abi_bytes;
+// mod abi_bytes;
 mod abi_token;
